requirements.txt
ddgs==9.5.5
fastapi==0.116.2
numpy==2.3.3
opencv_python==4.12.0.88
Pillow==11.3.0
pydantic==2.11.9
Requests==2.32.5
torch==2.8.0
torchvision==0.23.0
